requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
-e git://github.com/arthurio/django@40793b940ded22a175fbe8ac084f2d8f77ae27b3#egg=Django
django-extensions==2.1.2
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pudb==2018.1
Pygments==2.2.0
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urwid==2.0.1
wcwidth==0.1.7
